refactor: revamp states and url
