+++ packages/manager/modules/cda/src/details/cda-details.html
@@ -18,13 +18,13 @@
                 <span data-translate="cda_tab_user"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-ip.cda-ip-list"
+                state="cda.dashboard.cda-ip"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_ip"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-pool.cda-pool-list"
+                state="cda.dashboard.cda-pool"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_pool"></span>
