+++ packages/manager/modules/cda/src/ip/cda-ip-list.html
@@ -1,4 +1,4 @@
-<section>
+<section data-ui-view="cdaIpContent">
     <header>
         <p data-translate="cda_ip_ip_definition"></p>
         <p data-translate="cda_ip_permission_definition"></p>
