--- packages/manager/modules/cda/src/ip/cda-ip.html
@@ -1,1 +1,0 @@
-<div data-ui-view="cdaIpContent"></div>
