+++ packages/manager/modules/cda/src/ip/index.js
@@ -1,14 +1,13 @@
 import angular from 'angular';
 
-import routing from './cda-ip';
-import cdaUserList from './list';
+import routing from './routing';
 import addController from './add/cda-ip-add.controller';
 import deleteController from './delete/cda-ip-delete.controller';
 
 const moduleName = 'ovhManagerCdaIp';
 
 angular
-  .module(moduleName, [cdaUserList])
+  .module(moduleName, [])
   .config(routing)
   .controller('CdaIpAddCtrl', addController)
   .controller('CdaIpDeleteCtrl', deleteController)
