--- packages/manager/modules/cda/src/ip/list/cda-ip-list.js
@@ -1,19 +1,0 @@
-import controller from './cda-ip-list.controller';
-import template from './cda-ip-list.html';
-
-export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-ip.cda-ip-list', {
-    url: '/list',
-    views: {
-      cdaIpContent: {
-        template,
-        controller,
-        controllerAs: 'CdaIpListCtrl',
-      },
-    },
-    translations: {
-      format: 'json',
-      value: ['.'],
-    },
-  });
-};
