--- packages/manager/modules/cda/src/ip/list/index.js
@@ -1,12 +1,0 @@
-import angular from 'angular';
-
-import routing from './cda-ip-list';
-
-const moduleName = 'ovhManagerCdaIpList';
-
-angular
-  .module(moduleName, [])
-  .config(routing)
-  .run(/* @ngTranslationsInject:json ./translations */);
-
-export default moduleName;
