--- packages/manager/modules/cda/src/ip/list/translations/Messages_it_IT.json
@@ -1,18 +1,0 @@
-{
-  "cda_ip_ip_definition": "Crea qui la tua lista di controllo degli accessi IP.",
-  "cda_ip_permission_definition": "Il tuo cluster accetterà solo il traffico di rete che corrisponde agli accessi creati.",
-  "cda_ip_add_button_text": "Aggiungi una lista di controllo degli accessi",
-  "cda_ip_delete_button_text": "Elimina",
-  "cda_ip_table_col_network": "IP",
-  "cda_ip_table_col_actions": "Azioni",
-  "cda_ip_add_title": "Aggiungi una lista di controllo degli accessi",
-  "cda_ip_add_ip_label": "IP con maschera di sottorete",
-  "cda_ip_add_format": "Formato: 192.168.1.0/24",
-  "cda_ip_add_success": "Lista di controllo degli accessi aggiunta correttamente.",
-  "cda_ip_add_ip_required": "È necessario indicare l'IP con maschera di sottorete.",
-  "cda_ip_delete_title": "Elimina una lista di controllo degli accessi",
-  "cda_ip_delete_ip_label": "IP con maschera di sottorete",
-  "cda_ip_delete_confirm": "Vuoi davvero eliminare questa lista di controllo degli accessi?",
-  "cda_ip_delete_success": "Lista di controllo degli accessi eliminata correttamente.",
-  "cda_ip_table_empty": "Nessuna lista di controllo degli accessi."
-}
