--- packages/manager/modules/cda/src/ip/list/translations/Messages_pl_PL.json
@@ -1,18 +1,0 @@
-{
-  "cda_ip_ip_definition": "Tutaj możesz utworzyć reguły kontroli dostępu IP.",
-  "cda_ip_permission_definition": "Twój klaster będzie akceptował wyłącznie ruch sieciowy połączony z utworzonymi dostępami.",
-  "cda_ip_add_button_text": "Dodaj kontrolę dostępu",
-  "cda_ip_delete_button_text": "Usuń",
-  "cda_ip_table_col_network": "IP",
-  "cda_ip_table_col_actions": "Operacje",
-  "cda_ip_add_title": "Dodaj kontrolę dostępu",
-  "cda_ip_add_ip_label": "IP z maską podsieci",
-  "cda_ip_add_format": "Format: 192.168.1.0/24",
-  "cda_ip_add_success": "Kontrola dostępu została dodana.",
-  "cda_ip_add_ip_required": "Wymagane jest IP z maską podsieci.",
-  "cda_ip_delete_title": "Usuń kontrolę dostępu",
-  "cda_ip_delete_ip_label": "IP z maską podsieci",
-  "cda_ip_delete_confirm": "Czy chcesz usunąć tę kontrolę dostępu?",
-  "cda_ip_delete_success": "Kontrola dostępu została usunięta.",
-  "cda_ip_table_empty": "Brak kontroli dostępu."
-}
