+++ packages/manager/modules/cda/src/ip/translations/Messages_de_DE.json
@@ -1,1 +1,18 @@
-{}
+{
+  "cda_ip_ip_definition": "Hier können Sie IP-Zugriffskontrollen einrichten. ",
+  "cda_ip_permission_definition": "Ihr Cluster wird nur Netzwerktraffic entsprechend den eingerichteten Zugriffsbeschränkungen zulassen. ",
+  "cda_ip_add_button_text": "Zugriffskontrolle hinzufügen ",
+  "cda_ip_delete_button_text": "Löschen",
+  "cda_ip_table_col_network": "IP",
+  "cda_ip_table_col_actions": "Aktionen",
+  "cda_ip_add_title": "Zugriffskontrolle hinzufügen ",
+  "cda_ip_add_ip_label": "IP mit Subnetzmaske ",
+  "cda_ip_add_format": "Format: 192.168.1.0/24 ",
+  "cda_ip_add_success": "Zugriffskontrolle erfolgreich hinzugefügt. ",
+  "cda_ip_add_ip_required": "IP mit Subnetzmaske erforderlich. ",
+  "cda_ip_delete_title": "Zugriffskontrolle löschen ",
+  "cda_ip_delete_ip_label": "IP mit Subnetzmaske ",
+  "cda_ip_delete_confirm": "Wollen Sie diese Zugriffskontrolle wirklich löschen? ",
+  "cda_ip_delete_success": "Zugriffskontrolle erfolgreich gelöscht. ",
+  "cda_ip_table_empty": "Keine Zugriffskontrollen eingerichtet. "
+}
