+++ packages/manager/modules/cda/src/ip/translations/Messages_en_GB.json
@@ -1,1 +1,18 @@
-{}
+{
+  "cda_ip_ip_definition": "You can create your IP access control list here.",
+  "cda_ip_permission_definition": "Your Cloud Disk Array cluster will accept only such network traffic which matches created Access Controls.",
+  "cda_ip_add_button_text": "Add IP Access Control",
+  "cda_ip_delete_button_text": "Delete",
+  "cda_ip_table_col_network": "IP",
+  "cda_ip_table_col_actions": "Actions",
+  "cda_ip_add_title": "Add IP Access Control",
+  "cda_ip_add_ip_label": "IP with netmask",
+  "cda_ip_add_format": "Format: 192.168.1.0/24",
+  "cda_ip_add_success": "The new ACL has been created.",
+  "cda_ip_add_ip_required": "IP with netmask is required.",
+  "cda_ip_delete_title": "Delete ACL",
+  "cda_ip_delete_ip_label": "IP with netmask",
+  "cda_ip_delete_confirm": "Are you sure you want do delete this ACL?",
+  "cda_ip_delete_success": "The ACL has been deleted.",
+  "cda_ip_table_empty": "No ACL available."
+}
