+++ packages/manager/modules/cda/src/ip/translations/Messages_fr_CA.json
@@ -1,1 +1,18 @@
-{}
+{
+  "cda_ip_ip_definition": "Vous pouvez créer vos contrôles d'accès IP ici.",
+  "cda_ip_permission_definition": "Votre cluster acceptera seulement le trafic réseau concordant avec les accès créés.",
+  "cda_ip_add_button_text": "Ajouter un contrôle d'accès",
+  "cda_ip_delete_button_text": "Supprimer",
+  "cda_ip_table_col_network": "IP",
+  "cda_ip_table_col_actions": "Actions",
+  "cda_ip_add_title": "Ajouter un contrôle d'accès",
+  "cda_ip_add_ip_label": "IP avec masque de sous-réseau",
+  "cda_ip_add_format": "Format : 192.168.1.0/24",
+  "cda_ip_add_success": "Controle d'accès ajouté avec succès.",
+  "cda_ip_add_ip_required": "L'IP avec masque de sous-réseau est requis.",
+  "cda_ip_delete_title": "Supprimer un contrôle d'accès",
+  "cda_ip_delete_ip_label": "IP avec masque de sous-réseau",
+  "cda_ip_delete_confirm": "Êtes-vous certain de vouloir supprimer ce contrôle d'accès?",
+  "cda_ip_delete_success": "Controle d'accès supprimé avec succès.",
+  "cda_ip_table_empty": "Aucun contrôle d'accès."
+}
