+++ packages/manager/modules/cda/src/pool/cda-pool-list.html
@@ -1,4 +1,4 @@
-<section>
+<section data-ui-view="cdaPoolContent">
     <header>
         <p data-translate="cda_pool_pool_definition"></p>
         <p data-translate="cda_pool_permission_definition"></p>
