--- packages/manager/modules/cda/src/pool/cda-pool.html
@@ -1,1 +1,0 @@
-<div data-ui-view="cdaPoolContent"></div>
