--- packages/manager/modules/cda/src/pool/cda-pool.js
@@ -1,17 +1,0 @@
-import template from './cda-pool.html';
-
-export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-pool', {
-    url: '/pool',
-    views: {
-      cdaDetailsTab: {
-        template,
-        abstract: true,
-      },
-    },
-    translations: {
-      format: 'json',
-      value: ['.'],
-    },
-  });
-};
