+++ packages/manager/modules/cda/src/pool/index.js
@@ -1,13 +1,16 @@
 import angular from 'angular';
 
-import routing from './cda-pool';
-import cdaPoolList from './list';
+import routing from './routing';
+import addController from './add/cda-pool-add.controller';
+import deleteController from './delete/cda-pool-delete.controller';
 
 const moduleName = 'ovhManagerCdaPool';
 
 angular
-  .module(moduleName, [cdaPoolList])
+  .module(moduleName, [])
   .config(routing)
+  .controller('CdaPoolAddCtrl', addController)
+  .controller('CdaPoolDeleteCtrl', deleteController)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
