--- packages/manager/modules/cda/src/pool/list/translations/Messages_de_DE.json
@@ -1,23 +1,0 @@
-{
-  "cda_pool_pool_definition": "Hier können Sie Ihre Pools erstellen. ",
-  "cda_pool_permission_definition": "Pools sind \"Container\" für Ihre Daten. Der Zugriff auf diese Pools kann für jeden Benutzer einzeln festgelegt werden. ",
-  "cda_pool_add_button_text": "Pool hinzufügen ",
-  "cda_pool_delete_button_text": "Löschen",
-  "cda_pool_table_col_poolname": "Name",
-  "cda_pool_table_col_pooltype": "Typ",
-  "cda_pool_table_col_replica": "Anzahl der Replikate ",
-  "cda_pool_table_col_minreplica": "Mindestzahl aktiver Replikate ",
-  "cda_pool_table_col_backup": "Backup Pool ",
-  "cda_pool_table_col_actions": "Aktionen",
-  "cda_pool_add_title": "Pool hinzufügen ",
-  "cda_pool_add_poolname_label": "Name",
-  "cda_pool_add_poolname_required": "Bitte geben Sie einen Namen an. ",
-  "cda_pool_add_poolname_pattern": "Der Name darf keine Leerzeichen enthalten. ",
-  "cda_pool_add_poolname_maxlength": "Der Name darf maximal {{ maxlength }} Zeichen umfassen. ",
-  "cda_pool_add_success": "Pool erfolgreich hinzugefügt. ",
-  "cda_pool_delete_title": "Pool löschen ",
-  "cda_pool_delete_poolname_label": "Name",
-  "cda_pool_delete_confirm": "Wollen Sie diesen Pool wirklich löschen? ",
-  "cda_pool_delete_success": "Pool erfolgreich gelöscht. ",
-  "cda_pool_table_empty": "Keine Pools vorhanden. "
-}
