--- packages/manager/modules/cda/src/pool/list/translations/Messages_lt_LT.json
@@ -1,23 +1,0 @@
-{
-  "cda_pool_pool_definition": "Here is where you can create your pools.",
-  "cda_pool_permission_definition": "Pools are \"containers\" for your data. The user can manage access to the pools. ",
-  "cda_pool_add_button_text": "Add a pool",
-  "cda_pool_delete_button_text": "Delete",
-  "cda_pool_table_col_poolname": "Name",
-  "cda_pool_table_col_pooltype": "Type",
-  "cda_pool_table_col_replica": "Number of replicas",
-  "cda_pool_table_col_minreplica": "Minimum active replicas",
-  "cda_pool_table_col_backup": "Backup pool",
-  "cda_pool_table_col_actions": "Actions",
-  "cda_pool_add_title": "Add a pool",
-  "cda_pool_add_poolname_label": "Name",
-  "cda_pool_add_poolname_required": "You must name your cluster.",
-  "cda_pool_add_poolname_pattern": "The name cannot contain spaces.",
-  "cda_pool_add_poolname_maxlength": "The name of your cluster must not exceed {{ maxlength }} characters.",
-  "cda_pool_add_success": "Pool successfully added.",
-  "cda_pool_delete_title": "Delete pool",
-  "cda_pool_delete_poolname_label": "Name:",
-  "cda_pool_delete_confirm": "Are you sure you want to delete this pool?",
-  "cda_pool_delete_success": "Pool successfully deleted",
-  "cda_pool_table_empty": "No pools"
-}
