+++ packages/manager/modules/cda/src/pool/translations/Messages_fr_FR.json
@@ -1,1 +1,23 @@
-{}
+{
+  "cda_pool_pool_definition": "Vous pouvez créer vos pools ici.",
+  "cda_pool_permission_definition": "Les pools sont des \"containers\" pour vos données. L'accès aux pools peut être contrôlé par utilisateur.",
+  "cda_pool_add_button_text": "Ajouter un pool",
+  "cda_pool_delete_button_text": "Supprimer",
+  "cda_pool_table_col_poolname": "Nom",
+  "cda_pool_table_col_pooltype": "Type",
+  "cda_pool_table_col_replica": "Nombre de répliques",
+  "cda_pool_table_col_minreplica": "Nombre minimal de répliques actives",
+  "cda_pool_table_col_backup": "Backup pool",
+  "cda_pool_table_col_actions": "Actions",
+  "cda_pool_add_title": "Ajouter un pool",
+  "cda_pool_add_poolname_label": "Nom",
+  "cda_pool_add_poolname_required": "Le nom est requis.",
+  "cda_pool_add_poolname_pattern": "Le nom ne peut contenir d'espaces.",
+  "cda_pool_add_poolname_maxlength": "Le nom doit avoir une longueur maximale de {{ maxlength }} caractères.",
+  "cda_pool_add_success": "Pool ajouté avec succès.",
+  "cda_pool_delete_title": "Supprimer un pool",
+  "cda_pool_delete_poolname_label": "Nom :",
+  "cda_pool_delete_confirm": "Êtes-vous certain de vouloir supprimer ce pool?",
+  "cda_pool_delete_success": "Pool supprimé avec succès.",
+  "cda_pool_table_empty": "Aucun pool."
+}
