+++ packages/manager/modules/cda/src/pool/translations/Messages_it_IT.json
@@ -1,1 +1,23 @@
-{}
+{
+  "cda_pool_pool_definition": "Crea qui i tuoi pool.",
+  "cda_pool_permission_definition": "I pool sono \"container\" per i dati, il cui accesso viene gestito individualmente dagli utenti.",
+  "cda_pool_add_button_text": "Aggiungi un pool",
+  "cda_pool_delete_button_text": "Elimina",
+  "cda_pool_table_col_poolname": "Nome",
+  "cda_pool_table_col_pooltype": "Tipo",
+  "cda_pool_table_col_replica": "Numero di repliche",
+  "cda_pool_table_col_minreplica": "Numero minimo di repliche attive",
+  "cda_pool_table_col_backup": "Backup pool",
+  "cda_pool_table_col_actions": "Azioni",
+  "cda_pool_add_title": "Aggiungi un pool",
+  "cda_pool_add_poolname_label": "Nome",
+  "cda_pool_add_poolname_required": "È necessario inserire il nome.",
+  "cda_pool_add_poolname_pattern": "Il nome non può contenere spazi.",
+  "cda_pool_add_poolname_maxlength": "Il nome deve avere una lunghezza massima di {{ maxlength }} caratteri.",
+  "cda_pool_add_success": "Pool aggiunto correttamente.",
+  "cda_pool_delete_title": "Elimina un pool",
+  "cda_pool_delete_poolname_label": "Nome",
+  "cda_pool_delete_confirm": "Vuoi davvero eliminare questo pool?",
+  "cda_pool_delete_success": "Pool eliminato correttamente.",
+  "cda_pool_table_empty": "Nessun pool."
+}
