+++ packages/manager/modules/cda/src/pool/translations/Messages_pl_PL.json
@@ -1,1 +1,23 @@
-{}
+{
+  "cda_pool_pool_definition": "Tutaj może dodawać zakresy.",
+  "cda_pool_permission_definition": "Zakresy to \"kontenery\" dla danych. Dostęp do zakresów może być kontrolowany przez użytkownika. ",
+  "cda_pool_add_button_text": "Dodaj zakres",
+  "cda_pool_delete_button_text": "Usuń",
+  "cda_pool_table_col_poolname": "Nazwa",
+  "cda_pool_table_col_pooltype": "Typ",
+  "cda_pool_table_col_replica": "Liczba kopii",
+  "cda_pool_table_col_minreplica": "Minimalna liczba aktywnych kopii",
+  "cda_pool_table_col_backup": "Backup zakresu",
+  "cda_pool_table_col_actions": "Operacje",
+  "cda_pool_add_title": "Dodaj zakres",
+  "cda_pool_add_poolname_label": "Nazwa",
+  "cda_pool_add_poolname_required": "Nazwa jest wymagana.",
+  "cda_pool_add_poolname_pattern": "Nazwa nie może zawierać spacji.",
+  "cda_pool_add_poolname_maxlength": "Maksymalna liczba znaków w nazwie: {{ maxlength }} ",
+  "cda_pool_add_success": "Zakres został dodany.",
+  "cda_pool_delete_title": "Usuń zakres",
+  "cda_pool_delete_poolname_label": "Nazwa",
+  "cda_pool_delete_confirm": "Czy chcesz usunąć ten zakres?",
+  "cda_pool_delete_success": "Zakres został usunięty.",
+  "cda_pool_table_empty": "Brak zakresów."
+}
