+++ packages/manager/modules/cda/src/pool/translations/Messages_pt_PT.json
@@ -1,1 +1,23 @@
-{}
+{
+  "cda_pool_pool_definition": "Poderá criar as suas pools aqui.",
+  "cda_pool_permission_definition": "As pools são \"containers\" para os seus dados. é possível controlar o acesso às pools por utilizador.",
+  "cda_pool_add_button_text": "Adicionar uma pool",
+  "cda_pool_delete_button_text": "Eliminar",
+  "cda_pool_table_col_poolname": "Apelido",
+  "cda_pool_table_col_pooltype": "Tipo",
+  "cda_pool_table_col_replica": "Número de réplicas",
+  "cda_pool_table_col_minreplica": "Número mínimo de réplicas ativas",
+  "cda_pool_table_col_backup": "Pool de backup",
+  "cda_pool_table_col_actions": "Ações",
+  "cda_pool_add_title": "Adicionar uma pool",
+  "cda_pool_add_poolname_label": "Apelido",
+  "cda_pool_add_poolname_required": "O nome é obrigatório.",
+  "cda_pool_add_poolname_pattern": "O nome não pode conter espaços.",
+  "cda_pool_add_poolname_maxlength": "O nome deve ter um comprimento máximo de {{ maxlength }} caracteres.",
+  "cda_pool_add_success": "Pool adicionado com sucesso.",
+  "cda_pool_delete_title": "Eliminar uma pool",
+  "cda_pool_delete_poolname_label": "Nome",
+  "cda_pool_delete_confirm": "Têm a certeza de que deseja eliminar a pool?",
+  "cda_pool_delete_success": "Pool eliminada com sucesso.",
+  "cda_pool_table_empty": "Sem pool."
+}
